requirements.txt
Fabric==1.10.1
fabtools==0.19.0
pycurl==7.19.5.1
python-geohash==0.8.5
simplejson==3.7.3
tornado==4.2
tornado-cors==0.6.0
tornado-redis==2.4.18
requests
